Just unify the mark
