+++ src/proxyfactory.rs
@@ -67,11 +67,12 @@ impl ProxyFactory {
     }
 
     pub async fn new_proxies(&self) -> Result<ProxyResult, Error> {
-        let factory = if self.config.testing_dedicated_mark != 0 {
-            new_mark_socket_factory(self.config.testing_dedicated_mark)
-        } else {
-            Arc::new(crate::proxy::DefaultSocketFactory)
-        };
+        let factory: Arc<dyn crate::proxy::SocketFactory + Send + Sync> =
+            if let Some(mark) = self.config.packet_mark {
+                Arc::new(crate::proxy::MarkSocketFactory(mark))
+            } else {
+                Arc::new(crate::proxy::DefaultSocketFactory)
+            };
         self.new_proxies_from_factory(None, None, factory).await
     }
 
@@ -132,14 +133,4 @@ pub struct ProxyResult {
     pub proxy: Option<Proxy>,
     pub dns_proxy: Option<dns::Server>,
     pub connection_manager: Option<ConnectionManager>,
-}
-
-#[cfg(feature = "testing")]
-fn new_mark_socket_factory(mark: u32) -> Arc<dyn crate::proxy::SocketFactory + Send + Sync> {
-    Arc::new(crate::test_helpers::MarkSocketFactory(mark))
-}
-
-#[cfg(not(feature = "testing"))]
-fn new_mark_socket_factory(_mark: u32) -> Arc<dyn crate::proxy::SocketFactory + Send + Sync> {
-    unimplemented!("testing_dedicated_mark requires --features testing")
 }
